Added game over state and formed the basis for level 1's concept
Added game over screen that pauses all inputs from being taken. along with a restart button to reload the current scene and exit button which will bring the person playing back to the game's main menu. Also created the concept of the first level of the game. As a note the enemies are placeholders in the level for now as they serve no purpose besides looking nice for the concept.