Added a death check
If the player falls off the stage and reaches a y value of -10, the player object is deleted and the dead variable is set to 1. If the dead variable is set to one the game over state is then run where the timer is set to 0 allowing the player to restart or exit.