Big changes to the game
Added more code to the player controller script to make a less floaty jump. Also added enemy and spike collisions where when touched the player is destroyed and the UI for restart and quit appear. Made patrolling script for the enemies to move along the platforms they are positioned on. Created thinner platforms and changed the scale of the player for more enjoyable traversal between platforms.