Big Changes
Added a new bunny sprite that is helpful for the player to allow for higher jumps. Finished all level concepts while playtesting while making sure they are actually doable with multiple solutions to each level. Made the how to play menu and the main menu page with buttons to start the game, view how to play, and quit. This is the game's first beta. Work that still needs to be completed:
Add animation to Turtleboy, add game sounds, change the font for the text and buttons, create a background for the game. Possibly create more levels if time permits.